New module mrtrix3.utils
This module can be a collection point for functions or classes that are of utility but don't sensibly fit into any of the other named modules.
