+++ lib/mrtrix3/_5ttgen/fsl.py
@@ -39,9 +39,9 @@ def get_inputs(): #pylint: disable=unused-variable
 def execute(): #pylint: disable=unused-variable
   import math, os
   from distutils.spawn import find_executable
-  from mrtrix3 import app, fsl, image, is_windows, MRtrixError, path, run
+  from mrtrix3 import app, fsl, image, MRtrixError, path, run, utils
 
-  if is_windows():
+  if utils.is_windows():
     raise MRtrixError('\'fsl\' algorithm of 5ttgen script cannot be run on Windows: FSL not available on Windows')
 
   fsl_path = os.environ.get('FSLDIR', '')
